requirements.txt
altgraph==0.17.2
importlib-metadata==4.11.3
macholib==1.16
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.4
PyJWT==2.4.0
PyYAML==6.0
typing_extensions==4.2.0
zipp==3.8.0
